:recycle: refactor: Footer, Header 이미지 형식 변경
